+++ xenonpy/inverse/iqspr/__init__.py
@@ -2,4 +2,8 @@
 #  Use of this source code is governed by a BSD-style
 #  license that can be found in the LICENSE file.
 
+__all__ = ['IQSPR', 'BayesianRidgeEstimator', 'NGram']
+
+from .estimator import BayesianRidgeEstimator
 from .iqspr import IQSPR
+from .modifier import NGram
