+++ xenonpy/model/__init__.py
@@ -2,5 +2,7 @@
 #  Use of this source code is governed by a BSD-style
 #  license that can be found in the LICENSE file.
 
+__all__ = ['nn']
+
 from . import extern
 from . import nn
