+++ xenonpy/model/nn/__init__.py
@@ -2,8 +2,10 @@
 #  Use of this source code is governed by a BSD-style
 #  license that can be found in the LICENSE file.
 
+__all__ = ['wrap', 'Checker', 'Layer1d', 'RegressionRunner', 'persist', 'BaseRunner']
+
 from . import wrap
-from .checker import *
+from .checker import Checker
 from .layer import Layer1d
 from .model_maker import Generator1d
-from .runner import *
+from .runner import BaseRunner, RegressionRunner, persist
